refactor(connector): move EmbassyUart to io module and update dependencies
