+++ examples/embassy-knx-connector-demo/Cargo.toml
@@ -17,6 +17,8 @@ aimdb-core = { path = "../../aimdb-core", default-features = false, features = [
 ] }
 aimdb-embassy-adapter = { path = "../../aimdb-embassy-adapter", default-features = false, features = [
     "embassy-runtime",
+    # `EmbassyUart`, the byte source the serial connector frames.
+    "connector-io",
 ] }
 aimdb-knx-connector = { path = "../../aimdb-knx-connector", default-features = false, features = [
     "embassy-runtime",
@@ -24,7 +26,7 @@ aimdb-knx-connector = { path = "../../aimdb-knx-connector", default-features = f
 ] }
 # Serial remote-access server — serves this db's records over a UART (ST-LINK VCP)
 # so a host can read them with `aimdb --features transport-serial --connect serial://…`.
-# Its `embassy-runtime` feature also turns on `aimdb-core/remote`.
+# Its `connector` feature also turns on `aimdb-core/remote`.
 aimdb-serial-connector = { path = "../../aimdb-serial-connector", default-features = false, features = [
     "connector",
     "defmt",
